Move grid Row and tests into submodules
This is part of some cleanup for the grid module as a whole.
